refactor(recnet-api-model): reorganize api model by features
