+++ libs/recnet-api-model/src/index.ts
@@ -1,4 +1,6 @@
 export * from "./lib/model";
-export * from "./lib/api/admin";
 export * from "./lib/api/rec";
 export * from "./lib/api/user";
+export * from "./lib/api/stat";
+export * from "./lib/api/invite-code";
+export * from "./lib/api/article";
