+++ libs/recnet-api-model/src/lib/api/rec.ts
@@ -80,14 +80,3 @@ export const patchRecsUpcomingResponseSchema = z.object({
 export type PatchRecsUpcomingResponse = z.infer<
   typeof patchRecsUpcomingResponseSchema
 >;
-
-// GET /articles
-export const getArticlesParamsSchema = z.object({
-  link: z.string().url(),
-});
-export type GetArticlesParams = z.infer<typeof getArticlesParamsSchema>;
-
-export const getArticlesResponseSchema = z.object({
-  article: articleSchema.nullable(),
-});
-export type GetArticlesResponse = z.infer<typeof getArticlesResponseSchema>;
