+++ libs/recnet-api-model/src/lib/api/stat.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+// GET /stats
+export const getStatsResponseSchema = z.object({
+  numUsers: z.number(),
+  numRecs: z.number(),
+  numRecsOverTime: z.record(z.string(), z.number()),
+  numUpcomingRecs: z.number(),
+});
+export type GetStatsResponse = z.infer<typeof getStatsResponseSchema>;
